+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to
 
 
 ## [unreleased]
+### Added
+- added ctrl + a (select all) and ctrl + c (copy) to SqlDialog
+
 ### Fixed
+- Instruction rendering issues
 - FindAllResults scroll issues
 - Temporarily fixed highlights of text in Spans by stripping out new line chars.